pyproject.toml
[tool.poetry]
name = "netplanner"
version = "0.11.2"
description = "Mimir the Netplanner is the ground of all network wisdöm"
authors = ["Marcel Fest <marcel.fest@telekom.de>", "Christopher Dziomba <christopher.dziomba@telekom.de>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
dacite = "^1.6.0"
PyYAML = "^6.0"
fqdn = "^1.5.1"
Jinja2 = "^3.1.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
mypy = "^0.910"
types-PyYAML = "^6.0"
pytest = "^7.1.1"
pyoxidizer = "^0.20.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
netplanner = 'netplanner.__main__:main'
